pyproject.toml
[project]
name = "unplanned_travel_planner"
version = "0.1.0"
description = "unplanned_travel_planner using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "crewai>=0.201.0,<1.0.0",
    "google-generativeai>=0.7.0",
    "langchain-google-genai>=2.0.0",
    "langchain>=0.2.0",
    "python-dotenv>=1.0.1"
]

[project.scripts]
unplanned_travel_planner = "unplanned_travel_planner.main:run"
run_crew = "unplanned_travel_planner.main:run"
train = "unplanned_travel_planner.main:train"
replay = "unplanned_travel_planner.main:replay"
test = "unplanned_travel_planner.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
